Modularized all existing code
