--- .sample.env
@@ -1,7 +1,0 @@
-APP_API_KEY="YOUR_STUFF_HERE"
-APP_AUTH_DOMAIN="YOUR_STUFF_HERE"
-APP_DATABASE_URL="YOUR_STUFF_HERE"
-APP_PROJECT_ID="YOUR_STUFF_HERE"
-APP_STORAGE_BUCKET="YOUR_STUFF_HERE"
-APP_APP_ID="YOUR_STUFF_HERE"
-APP_MEASUREMENT_ID="YOUR_STUFF_HERE"
