+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "webpack-5.0-template",
+  "name": "lab-error-clean-up",
   "version": "1.0.0",
   "description": "",
   "main": "src/scripts/main.js",
@@ -9,7 +9,7 @@
     "build": "webpack --mode production"
   },
   "keywords": [],
-  "author": "",
+  "author": "Massacre-Saint",
   "license": "ISC",
   "devDependencies": {
     "css-loader": "^6.2.0",
@@ -21,7 +21,6 @@
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
     "mini-css-extract-plugin": "^2.2.2",
-    "node-sass": "^6.0.1",
     "sass-loader": "^12.1.0",
     "style-loader": "^3.2.1",
     "webpack": "^5.51.1",
@@ -34,6 +33,7 @@
     "axios": "^0.21.1",
     "bootstrap": "^5.1.0",
     "dotenv": "^10.0.0",
-    "firebase": "^8.2.0"
+    "firebase": "^8.2.0",
+    "sass": "^1.51.0"
   }
 }
