+++ src/scripts/components/header.js
@@ -1,0 +1,15 @@
+import renderToDOM from '../helpers/utilities';
+
+const header = () => {
+  const domString = `<div class="container">
+    <h1>Welcome to Hoggy Hogwarts Sorting Hat!</h1>
+    <p>
+      Hmm, difficult. VERY difficult. <br />Plenty of courage, I see.
+      <br />Not a bad mind, either. There's talent, oh yes. <br />And a
+      thirst to prove yourself. <br />But where to put you?
+    </p>
+  </div>`;
+
+  renderToDOM('#header-container', domString);
+};
+export default header;
