--- src/scripts/components/logoutButton.js
@@ -1,14 +1,0 @@
-import firebase from 'firebase/app';
-import 'firebase/auth';
-
-const signMeOut = () => {
-  firebase.auth().signOut();
-};
-
-const logoutButton = () => {
-  const domString = '<button id="google-auth" class="btn btn-danger">SIGNOUT</button>';
-  document.querySelector('#login-form-container').innerHTML = (domString);
-  document.querySelector('#google-auth').addEventListener('click', signMeOut);
-};
-
-export default logoutButton;
