+++ src/scripts/helpers/data/createId.js
@@ -1,0 +1,8 @@
+const createId = (array) => {
+  if (array.length) {
+    const idArray = array.map((el) => el.id);
+    return Math.max(...idArray) + 1;
+  } return 0;
+};
+
+export default createId;
