+++ src/scripts/helpers/data/sortingHat.js
@@ -1,0 +1,26 @@
+import { students } from './studentData';
+import houses from './housesData';
+import studentsOnDom from '../../components/studentCards';
+import createId from './createId';
+
+const sortStudent = (e) => {
+  e.preventDefault();
+  const sortingHat = houses[Math.floor(Math.random() * houses.length)];
+
+  if (e.target.id === 'sorting') {
+    const student = document.querySelector('#student-name');
+
+    // create the new student object
+    students.push({
+      id: createId(students),
+      name: student.value,
+      house: sortingHat.house,
+      crest: sortingHat.crest,
+    });
+
+    student.value = ''; // reset value of input
+    studentsOnDom('#students', students);
+  }
+};
+
+export default sortStudent;
