+++ src/scripts/helpers/data/studentData.js
@@ -1,0 +1,4 @@
+const students = [];
+const voldysArmy = [];
+
+export { students, voldysArmy };
