--- src/scripts/helpers/signOut.js
@@ -1,8 +1,0 @@
-import firebase from 'firebase/app';
-import 'firebase/auth';
-
-const signOut = () => {
-  firebase.auth().signOut();
-};
-
-export default signOut;
